refactor: inline card edit action
